% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaryScore.R
\docType{methods}
\name{computeBoundaryScores}
\alias{computeBoundaryScores}
\alias{computeBoundaryScores,ANY-method}
\title{Compute boundary scores}
\usage{
computeBoundaryScores(hic, sample, chrs, res, num_bins)

\S4method{computeBoundaryScores}{ANY}(hic, sample, chrs = paste0("chr",
  c(1:22)), res = "20000", num_bins = 5)
}
\arguments{
\item{hic}{A sparseHiC Object}

\item{sample}{The name of the sample (e.g. GM12878 or BT142-rep1)}

\item{chrs}{= paste0("chr", c(1:22))  Specify which chromosomes
to compute boundary scores for}

\item{res}{= "20000"}

\item{num_bins}{= "human" or "mouse" likely supported}
}
\value{
A data.frame of boundary scores
}
\description{
\code{computeBoundaryScores} returns a \code{data.frame} object
with boundary scores.
}
\examples{
# Need to make the example better
samp <- "IMR90"
# hesc <- getSampleFromBucket(samp, "dnalandscaper", c("chr1", "chr4"))

}

