% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\docType{methods}
\name{getChromosome}
\alias{getChromosome}
\alias{getChromosome,sparseHiCdatum,character-method}
\title{Grab a specific chromosome}
\usage{
getChromosome(obj, chr, a.list = TRUE)

\S4method{getChromosome}{sparseHiCdatum,character}(obj, chr, a.list = TRUE)
}
\arguments{
\item{obj}{A \code{sparseHiCdatum} object}

\item{chr}{Chromosome(s) desired}

\item{a.list}{= TRUE Return a list of matrices? If FALSE,
return another \code{sparseHiCdatum} object}
}
\value{
Returns object subsetted by resolution
}
\description{
\code{getChromosome} takes a \code{sparseHiCdatum} object
and returns a list of sparse Hi-C matrices. If a list is
returned, then the names are the resolutions. If the 
length of chr > 1 (multiple chromosomes), then
a list of lists will be returned if a.list is TRUE
}
\examples{
chr <- "chr21"
rds<-paste(system.file('rds',package='sparseHiC'),'hESCdatum1.rds',sep='/')
hESCdatum1 <- readRDS(rds)
r <- getChromosome(hESCdatum1,chr = chr, a.list = FALSE) 
}

