% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCPlotter.R
\docType{methods}
\name{getHiCcolor}
\alias{getHiCcolor}
\alias{getHiCcolor,ANY-method}
\title{Get a color ramp for plotting}
\usage{
getHiCcolor(p = character(0), name = character(0))

\S4method{getHiCcolor}{ANY}(p = character(0), name = character(0))
}
\arguments{
\item{p}{= Integer index of the plot color them listed below}

\item{name}{= Character name of the plot color theme listed below}
}
\description{
\code{getHiCcolor} returns a color ramp from one of the 
prespecified color options. This can either be accessed
from the name or the index (p) of the ramp
}
\examples{
color.choices <- list(
   "Pallet" = 1,
   "Viridian" = 2,
   "Pewter" = 3,
   "Cerulean" = 4,
   "Vermillion" = 5,
   "Lavendar" = 6,
   "Celadon" = 7,
   "Fuchsia" = 8,
   "Saffron" = 9,
   "Cinnabar" = 10,
   "Indigo" = 11,
   "Master" = 12,
   "Black and Blue" = 13,
   "Heat" = 14,
   "Topology" = 15,
   "Blue to Red" = 16)
   
}

