% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\docType{methods}
\name{getHiCMatrix}
\alias{getHiCMatrix}
\alias{getHiCMatrix,sparseHiCdatum-method}
\title{Grab a specific sparse Hi-C matrix}
\usage{
getHiCMatrix(obj, chr, res)

\S4method{getHiCMatrix}{sparseHiCdatum}(obj, chr, res)
}
\arguments{
\item{obj}{A \code{sparseHiCdatum} object}

\item{chr}{Chromosome desired}

\item{res}{Resolution desired}
}
\value{
Returns a Hi-C Matrix
}
\description{
\code{getHiCMatrix} takes a \code{sparseHiCdatum} object
and returns a single matrix. Thus, the length of chr
and res must be 1
}
\examples{
chr <- "chr7"
res <- "1000000"
}

