% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWS.R
\docType{methods}
\name{getSampleFromBucket}
\alias{getSampleFromBucket}
\alias{getSampleFromBucket,ANY-method}
\title{Import HiC Sample from AWS Bucket}
\usage{
getSampleFromBucket(sample, bucket, chr = NA, organism = "human")

\S4method{getSampleFromBucket}{ANY}(sample, bucket, chr = NA,
  organism = "human")
}
\arguments{
\item{sample}{The name of the sample (e.g. GM12878 or BT142-rep1)}

\item{bucket}{Name of the AWS bucket.}

\item{chr}{= NA Specify which chromosomes to pull down; if left with
NA, then by default, all chromosomes will be pulled down}

\item{organism}{= "human" or "mouse" likely supported}
}
\value{
A sparseHiCdatum object
}
\description{
\code{getSampleFromBucket} returns a \code{sparseHiCdatum} object
when the user specifies the sample name and the bucket that the object
is housed in. If the sample is split by chromosomes (e.g. GM12878 in
the DNAlandscapeR bucket), then the function collates those into one
object. However, the user can specify only certain chromosomes be 
imported using the chr paramter. As a side note, all AWS buckets are
lower case letters only, so that may be a source of error when trying
to import the bucket.
}
\examples{
# Import chromosome 4 from DNAlandscapeR bucket
samp <- "hESC1"
#hesc <- getSampleFromBucket(samp, "dnalandscaper2", c("chr1", "chr4"))
}

