% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWS.R
\docType{methods}
\name{getSamplesFromBucket}
\alias{getSamplesFromBucket}
\alias{getSamplesFromBucket,ANY-method}
\title{Import multiple Hi-C samples from AWS Bucket}
\usage{
getSamplesFromBucket(samples, bucket, chr = NA, organism = "human")

\S4method{getSamplesFromBucket}{ANY}(samples, bucket, chr = NA,
  organism = "human")
}
\arguments{
\item{samples}{The names of the samples}

\item{bucket}{Name of the AWS bucket.}

\item{chr}{= NA Specify which chromosomes to pull down; if left with
NA, then by default, all chromosomes will be pulled down}

\item{organism}{= "human" or "mouse" likely supported}
}
\value{
A sparseHiCdata object
}
\description{
\code{getSamplesFromBucket} returns a \code{sparseHiCdata} object
when the user specifies multiple sample names present in the bucket
}
\examples{
# Import chromosome 4 from DNAlandscapeR bucket
samples <- c("GM12878", "IMR90")
# two <- getSamplesFromBucket(samples, "dnalandscaper", c("chr1", "chr4"))
}

