% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hESCdata}
\alias{hESCdata}
\title{hESC Samples}
\format{A small loops object 
\describe{
  \item{HiCSamplesList}{hESC1 and hESC2 each with 1Mb with chr20-22 }
  \item{metaData}{Shows alignment to hg19}
}}
\source{
# hESC
hESCdatum1 <- readRDS("inst/rds/hESCdatum1.rds")
hESCdatum2 <- readRDS("inst/rds/hESCdatum2.rds")
hESCdata <- hESCdatum1 + hESCdatum2
saveRDS(hESCdata, "inst/rds/hESCdata.rds")
}
\usage{
hESCdata
}
\value{
A sparseHiCdata object for samples
}
\description{
A sparseHiCdata object for both the ESC replicates
}
\keyword{datasets}

