% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hESCdatum1}
\alias{hESCdatum1}
\title{hESC1 Sample}
\format{A small loops object 
\describe{
  \item{sampleName}{hESC1}
  \item{resolutionNamedList}{1Mb with chr20-22 }
  \item{metaData}{Shows alignment to hg19}
}}
\source{
# hESC1
matrix.files <- paste(system.file("extdata", package = "processedHiCdata"),
"HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_iced.matrix", sep = "/")
bed.files <- paste(system.file("extdata", package = "processedHiCdata"),
"HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_abs.bed", sep = "/")
e1 <- import.HiCPro(matrix.files, bed.files, resolutions, sampleName = "hESC1", genomeBuild)
hESCdatum1 <- getChromosome(e1, c("chr20", "chr21", "chr22"), FALSE)
saveRDS(hESCdatum1, "inst/rds/hESCdatum1.rds")
}
\value{
A sparseHiCdatum
}
\description{
A sparseHiCdatum object from the ESC replicate 1
}

