% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hESCdatum2}
\alias{hESCdatum2}
\title{hESC2 Sample}
\format{A small loops object 
\describe{
  \item{sampleName}{hESC2}
  \item{resolutionNamedList}{1Mb with chr20-22 }
  \item{metaData}{Shows alignment to hg19}
}}
\source{
# hESC2
matrix.files <- paste(system.file("extdata", package = "processedHiCdata"),
"HiC-Pro/hESC_Rep2/hESC_Rep2_1000000_iced.matrix", sep = "/")
bed.files <- paste(system.file("extdata", package = "processedHiCdata"),
"HiC-Pro/hESC_Rep2/hESC_Rep2_1000000_abs.bed", sep = "/")
e2 <- import.HiCPro(matrix.files, bed.files, resolutions, sampleName = "hESC1", genomeBuild)
hESCdatum2 <- getChromosome(e2, c("chr20", "chr21", "chr22"), FALSE)
saveRDS(hESCdatum2, "inst/rds/hESCdatum2.rds")
}
\usage{
hESCdatum2
}
\value{
A sparseHiCdatum
}
\description{
A sparseHiCdatum object from the ESC replicate 2
}
\keyword{datasets}

