% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toDNAlandscapeR.R
\docType{methods}
\name{HiCPro2DNAlandscapeR}
\alias{HiCPro2DNAlandscapeR}
\alias{HiCPro2DNAlandscapeR,ANY-method}
\title{Export DNAlandscapeR files straight from HiC Pro Output}
\usage{
HiCPro2DNAlandscapeR(matrix.files, bed.files, resolutions, sampleName,
  genomeBuild = NA, splitByChr = FALSE, out.dir = "",
  drop.chrom = c("chrY", "chrM"), manual.chr = NA, manual.dist = NA,
  tempFile = TRUE, BPPARAM = BiocParallel::bpparam())

\S4method{HiCPro2DNAlandscapeR}{ANY}(matrix.files, bed.files, resolutions,
  sampleName, genomeBuild = NA, splitByChr = FALSE, out.dir = "",
  drop.chrom = c("chrY", "chrM"), manual.chr = NA, manual.dist = NA,
  tempFile = TRUE, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{matrix.files}{Path to .matrix files from Hi-C Pro Output}

\item{bed.files}{Path to .bed files from Hi-C Pro Output}

\item{resolutions}{Character vector of the resolutions of the Hi-C output}

\item{sampleName}{Single string of Hi-C data sample that is imported. "-HiC" will 
automatically be appended.}

\item{genomeBuild}{= NA Can specify one of c("hg38", "hg19", "hg18", "mm10", "mm9", "mm8")
that are built-in options for Hi-C chromosomes and distances. If not of these options
are suitable, then use the \code{manual.chr} and \code{manual.dist} parameters}

\item{splitByChr}{= FALSE Split .rds files by chromosome for faster i/o in browser?}

\item{out.dir}{= "" The output directory to save the new files}

\item{drop.chrom}{= c("chrY", "chrM") Chromosomes dropped from compression}

\item{manual.chr}{= NA Specify a vector of chromosome names in the data}

\item{manual.dist}{= NA Specify a same length vector as manual.chr with the 
chromosomal distances corresponding to each element in the manual.chr}

\item{tempFile}{= TRUE Create a temporary file in Awk to make i/o faster
and more memory efficient.}

\item{BPPARAM}{= bpparam() Parameters to pass to bplapply}
}
\value{
Will produce two files in the specified out.dir
}
\description{
\code{HiCPro2DNAlandscapeR} takes the input files from
HiC Pro output and creates a .rds and a .resolutions.txt
set of files that can be visualized in
DNAlandscapeR (dnalandscaper.aryeelab.org). If
splitByChr is specified as TRUE, will create a folder with the sample 
name and an .rds of each chromosome for faster i/o. These split files
can only be linked via a bucket in DNAlandscapeR.
}
\examples{
#' # matrix.files <- paste(system.file("extdata", package = "processedHiCdata"),
# "HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_iced.matrix", sep = "/")
# bed.files <- paste(system.file("extdata", package = "processedHiCdata"),
# "HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_abs.bed", sep = "/")
genomeBuild <- "hg19"
resolutions <- "1000000"
sampleName = "hESC1"
#HiCPro2DNAlandscapeR(matrix.files, bed.files, resolutions, sampleName = "hESC1", genomeBuild)
#HiCPro2DNAlandscapeR(matrix.files, bed.files, resolutions,
#sampleName = "hESC1", genomeBuild, splitByChr = TRUE)
}

