% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicpro.R
\docType{methods}
\name{import.HiCPro.full}
\alias{import.HiCPro.full}
\alias{import.HiCPro.full,ANY-method}
\title{Compress all Hi-C Pro output data, including interchromosomal interactions}
\usage{
import.HiCPro.full(matrix.files, bed.files, resolutions, sampleName,
  genomeBuild = NA, drop.chrom = c("chrY", "chrM"), manual.chr = NA,
  manual.dist = NA, BPPARAM = BiocParallel::bpparam())

\S4method{import.HiCPro.full}{ANY}(matrix.files, bed.files, resolutions,
  sampleName, genomeBuild = NA, drop.chrom = c("chrY", "chrM"),
  manual.chr = NA, manual.dist = NA, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{matrix.files}{Paths to .matrix files from Hi-C Pro Output}

\item{bed.files}{Paths to .bed files from Hi-C Pro Output}

\item{resolutions}{Character vector of the resolutions of the Hi-C output}

\item{sampleName}{Single string of Hi-C data sample that is imported}

\item{genomeBuild}{= NA Can specify one of c("hg38", "hg19", "hg18", "mm10", "mm9", "mm8")
that are built-in options for Hi-C chromosomes and distances. If not of these options
are suitable, then use the \code{manual.chr} and \code{manual.dist} parameters}

\item{drop.chrom}{= c("chrY", "chrM")}

\item{manual.chr}{= NA Specify a vector of chromosome names in the data}

\item{manual.dist}{= NA Specify a same length vector as manual.chr with the 
chromosomal distances corresponding to each element in the manual.chr}

\item{BPPARAM}{= bpparam() Parameters to pass to bplapply}
}
\value{
sparseHiCdatum of all chromosomal interactions pairwise
}
\description{
\code{import.HiCPro.full} takes output data from HiC-Pro and compresses
it. All interactions will be preserved including interchromosomal. No values 
can be zeroed out like in the \code{import.HiCPro} function. Not recommended
function to utilize for memory disk concerns but we do support importing the
full data into this framework.
}
\examples{
# matrix.files <- paste(system.file("extdata", package = "processedHiCdata"),
# "HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_iced.matrix", sep = "/")
# bed.files <- paste(system.file("extdata", package = "processedHiCdata"),
# "HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_abs.bed", sep = "/")
genomeBuild <- "hg19"
resolutions <- "1000000"
sampleName = "hESC1"
# x <- import.HiCPro.full(matrix.file, bed.file, res, sampleName = "hESC1", genomeBuild)

}

