% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicpro.R
\docType{methods}
\name{import.HiCPro}
\alias{import.HiCPro}
\alias{import.HiCPro,ANY-method}
\title{Compress Hi-C Pro output data}
\usage{
import.HiCPro(matrix.files, bed.files, resolutions, sampleName,
  genomeBuild = NA, n = 40, drop.chrom = c("chrY", "chrM"),
  manual.chr = NA, manual.dist = NA, BPPARAM = BiocParallel::bpparam())

\S4method{import.HiCPro}{ANY}(matrix.files, bed.files, resolutions, sampleName,
  genomeBuild = NA, n = 40, drop.chrom = c("chrY", "chrM"),
  manual.chr = NA, manual.dist = NA, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{matrix.files}{Path to .matrix files from Hi-C Pro Output}

\item{bed.files}{Path to .bed files from Hi-C Pro Output}

\item{resolutions}{Character vector of the resolutions of the Hi-C output}

\item{sampleName}{Single string of Hi-C data sample that is imported}

\item{genomeBuild}{= NA Can specify one of c("hg38", "hg19", "hg18", "mm10", "mm9", "mm8")
that are built-in options for Hi-C chromosomes and distances. If not of these options
are suitable, then use the \code{manual.chr} and \code{manual.dist} parameters}

\item{n}{= 40 Number of off-diagonal rows to retain features to retain. 
If n = 0, retain the full data.}

\item{drop.chrom}{= c("chrY", "chrM") Chromosomes dropped from compression}

\item{manual.chr}{= NA Specify a vector of chromosome names in the data}

\item{manual.dist}{= NA Specify a same length vector as manual.chr with the 
chromosomal distances corresponding to each element in the manual.chr}

\item{BPPARAM}{= bpparam() Parameters to pass to bplapply}
}
\value{
sparseHiCdatum of all intrachromosomal interactions of n diagonals
}
\description{
\code{import.HiCPro} takes output data from HiC-Pro and compresses
it. Only intrachromosomal interactions will be preserved.
}
\details{
Supply three vectors of equivalent length; one specifies the location
of the .matrix files, one the location of the .bed files, and one
the resolution of the samples for each element in that order.
}
\examples{
# matrix.files <- paste(system.file("extdata", package = "processedHiCdata"),
# "HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_iced.matrix", sep = "/")
# bed.files <- paste(system.file("extdata", package = "processedHiCdata"),
# "HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_abs.bed", sep = "/")
genomeBuild <- "hg19"
resolutions <- "1000000"
sampleName = "hESC1"
# x <- import.HiCPro(matrix.file, bed.file, res, sampleName = "hESC1", genomeBuild)

}

