% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCPlotter.R
\docType{methods}
\name{plotHiC}
\alias{plotHiC}
\alias{plotHiC,ANY-method}
\title{Plot HiC samples}
\usage{
plotHiC(x, y, res, libraryNorm = TRUE, log2trans = TRUE, quantiles = c(5,
  95), color = getHiCcolor(16), missingco = "min", showLegend = TRUE,
  organism = "h", showGenes = FALSE)

\S4method{plotHiC}{ANY}(x, y, res, libraryNorm = TRUE, log2trans = TRUE,
  quantiles = c(5, 95), color = getHiCcolor(16), missingco = "min",
  showLegend = TRUE, organism = "h", showGenes = FALSE)
}
\arguments{
\item{x}{A sparseHiCdata object with samples to be visualized}

\item{y}{A GRanges object containing region of interest}

\item{res}{The resolution to be plotted}

\item{libraryNorm}{= TRUE When plotting multiple samples,
normalize for library complexity?}

\item{log2trans}{= TRUE log2transform the HiC data?}

\item{quantiles}{= c(5, 95). A vector of length 2 OR 
the boolean FALSE. Every value below the first element of the
vector (e.g. less than the 5th percentile) will be plotted as the 
minimum and everything bigger than the second element of the vector
(e.g. greater than the 95th percentile) will be plotted as the maximum.}

\item{color}{A colorRamp defined by the
getHiCcolor function. Default is 13 from the getHiCcolor function}

\item{missingco}{= "min" By default, missing pixels are plotted
with the minimum color, but users can specify other color options
so long as they exist in base R (e.g. 'red', 'black', etc.)}

\item{showLegend}{= TRUE display legend on plots?}

\item{organism}{= 'h';  'h' for human or 'm' for mouse supported
for viewing the gene annotation. Only matters if showGenes is TRUE}

\item{showGenes}{= FALSE; show gene annotation at bottom panel.}
}
\description{
\code{plotHiC} prints a base R image of the region specified for 
the sparseHiCdata supplied.
}
\examples{

library(GenomicRanges)
region <- GRanges(seqnames=c("22"),ranges=IRanges(start=c(0),end=c(9000000)))
rdsA<-paste(system.file('rds',package='sparseHiC'),'hESCdata.rds',sep='/')
hESCdata <- readRDS(rdsA)
region <- GRanges(seqnames=c("chr21"),ranges=IRanges(start=c(23000000),end=c(40000000)))
res <- "1000000"
a <- plotHiC(hESCdata, region, res, showGenes = FALSE)

}

