% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\docType{methods}
\name{+,sparseHiCdatum,sparseHiCdatum-method}
\alias{+,sparseHiCdatum,sparseHiCdatum-method}
\title{Combining together multiple Hi-C Samples}
\usage{
\S4method{+}{sparseHiCdatum,sparseHiCdatum}(e1, e2)
}
\arguments{
\item{e1}{A sparseHiCdatum object}

\item{e2}{A sparseHiCdatum object}
}
\value{
A sparseHiCdatum/data object depending on if joining multiple samples
}
\description{
First, we check if the resolutions are different but the sample
is the same. If so, then we append the new resolution, returning a
sparseHiCdatum object. If not, then we append it as a new sample.
The metaData handling is designed for functionality, not inclusiveness.
}

