% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWS.R
\docType{methods}
\name{samplesInBucket}
\alias{samplesInBucket}
\alias{samplesInBucket,character-method}
\title{List Available Hi-C Samples from an AWS Bucket}
\usage{
samplesInBucket(bucket)

\S4method{samplesInBucket}{character}(bucket)
}
\arguments{
\item{bucket}{name of the AWS bucket (all lower case)}
}
\description{
\code{samplesInBucket} returns a character vector of the samples
that can be imported from the specified bucket. Should also specify
the organism.
}
\examples{
# samplesInBucket("dnalandscaper")
}

