% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicpro.R
\docType{methods}
\name{sparseCompress.HiCPro.full}
\alias{sparseCompress.HiCPro.full}
\alias{sparseCompress.HiCPro.full,ANY-method}
\title{Compress all Hi-C Pro output data, including intrachromosomal interactions}
\usage{
sparseCompress.HiCPro.full(matrix.file, bed.file, res, genomeBuild = NA,
  out.pre, drop.chrom = c("chrY", "chrM"), dir.create = TRUE,
  compress = TRUE, manual.chr = NA, manual.dist = NA,
  BPPARAM = BiocParallel::bpparam())

\S4method{sparseCompress.HiCPro.full}{ANY}(matrix.file, bed.file, res,
  genomeBuild = NA, out.pre, drop.chrom = c("chrY", "chrM"),
  dir.create = TRUE, compress = TRUE, manual.chr = NA, manual.dist = NA,
  BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{matrix.file}{Path to .matrix file from Hi-C Pro Output}

\item{bed.file}{Path to .bed file from Hi-C Pro Output}

\item{res}{Character of the resolution of the Hi-C output}

\item{genomeBuild}{= NA Can specify one of c("hg38", "hg19", "hg18", "mm10", "mm9", "mm8")
that are built-in options for Hi-C chromosomes and distances. If not of these options
are suitable, then use the \code{manual.chr} and \code{manual.dist} parameters}

\item{out.pre}{File name prefix; will also be name of the new directory created. 
"_" + res will be appended to the output file name.}

\item{drop.chrom}{= c("chrY", "chrM")}

\item{dir.create}{Create directory for resulting .rds files?}

\item{compress}{Compress the resulting directory?}

\item{manual.chr}{= NA Specify a vector of chromosome names in the data}

\item{manual.dist}{= NA Specify a same length vector as manual.chr with the 
chromosomal distances corresponding to each element in the manual.chr}

\item{BPPARAM}{= bpparam() Parameters to pass to bplapply}
}
\value{
Pairwise .rds files will be saved to the disk of each chromosome interaction, 
including interchromosomal interactions
}
\description{
\code{sparseCompress.HiCPro.full} takes output data from HiC-Pro and compresses
it. All interactions will be preserved including intrachromosomal. No values 
can be zeroed out like in the \code{sparseCompress.HiCPro} function.
}
\details{
The option to return a list is not supported since the intrachromosomal data
will likely be larger and the number of pairwise chromosome-chromosome interactions
scales to O(n^2).
}
\examples{
# matrix.file <- paste(system.file("extdata", package = "processedHiCdata"),
# "HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_iced.matrix", sep = "/")
# bed.file <- paste(system.file("extdata", package = "processedHiCdata"),
# "HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_abs.bed", sep = "/")
genomeBuild <- "hg19"
res <- "1000000"
# x <- sparseCompress.HiCPro.full(matrix.file, bed.file, res, genomeBuild, "hESC_rep1")

}

