% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicpro.R
\docType{methods}
\name{sparseCompress.HiCPro}
\alias{sparseCompress.HiCPro}
\alias{sparseCompress.HiCPro,ANY-method}
\title{Compress Hi-C Pro output data}
\usage{
sparseCompress.HiCPro(matrix.file, bed.file, res, genomeBuild = NA, n = 40,
  out.pre = NA, drop.chrom = c("chrY", "chrM"), list = TRUE,
  save = FALSE, dir.create = FALSE, compress = FALSE, manual.chr = NA,
  manual.dist = NA, BPPARAM = BiocParallel::bpparam())

\S4method{sparseCompress.HiCPro}{ANY}(matrix.file, bed.file, res,
  genomeBuild = NA, n = 40, out.pre = NA, drop.chrom = c("chrY",
  "chrM"), list = TRUE, save = FALSE, dir.create = FALSE,
  compress = FALSE, manual.chr = NA, manual.dist = NA,
  BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{matrix.file}{Path to .matrix file from Hi-C Pro Output}

\item{bed.file}{Path to .bed file from Hi-C Pro Output}

\item{res}{Character of the resolution of the Hi-C output}

\item{genomeBuild}{= NA Can specify one of c("hg38", "hg19", "hg18", "mm10", "mm9", "mm8")
that are built-in options for Hi-C chromosomes and distances. If not of these options
are suitable, then use the \code{manual.chr} and \code{manual.dist} parameters}

\item{n}{Number of off-diagonal rows to retain features to retain. 
If n = 0, retain the full data.}

\item{out.pre}{= NA Prefix required if saving to disk}

\item{drop.chrom}{= c("chrY", "chrM") Chromosomes dropped from compression}

\item{list}{= TRUE If the user wants to split based on chromosome, must specify
that save is TRUE}

\item{save}{= FALSE Save to .rds in lieu of returning the value? Useful when splitting.}

\item{dir.create}{If list is FALSE and save is TRUE, this allows a new folder to be
created of the resulting .rds files}

\item{compress}{if dir.create is TRUE, compresses the directory (to .tgz) and removes
the raw data.}

\item{manual.chr}{= NA Specify a vector of chromosome names in the data}

\item{manual.dist}{= NA Specify a same length vector as manual.chr with the 
chromosomal distances corresponding to each element in the manual.chr}

\item{BPPARAM}{= bpparam() Parameters to pass to bplapply}
}
\value{
Either .rds files szved on the local disk or a list of sparse matrices
}
\description{
\code{sparseCompress.HiCPro} takes output data from HiC-Pro and compresses
it. Only intrachromosomal interactions will be preserved.
}
\details{
There are two main modes that can be indicated by the save logic variable.
If save = FALSE, this function will return a list of sparse matrices where
the names of the matrices correspond to the chromosomes. If save = TRUE,
one (if list = TRUE) or many (list = FALSE) .rds files will be created
of the compressed Hi-C interactions in sparse matrices.
}
\examples{
# matrix.file <- paste(system.file("extdata", package = "processedHiCdata"),
# "HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_iced.matrix", sep = "/")
# bed.file <- paste(system.file("extdata", package = "processedHiCdata"),
# "HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_abs.bed", sep = "/")
genomeBuild <- "hg19"
res <- "1000000"
# x <- sparseCompress.HiCPro(matrix.file, bed.file, res, genomeBuild)
# x <- sparseCompress.HiCPro(matrix.file, bed.file, res, genomeBuild, list = FALSE,
#          save = TRUE, out.pre = "LOL", dir.create = TRUE, compress = TRUE, n = 0)

}

