% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseHiC.R
\docType{package}
\name{sparseHiC}
\alias{sparseHiC}
\alias{sparseHiC-package}
\title{sparseHiC: Efficient parallelized compression of Hi-C Data}
\description{
The sparseHiC package contains functions to compress 
output data formats from common Hi-C pre-processing pipelines
and create more friendly data structures for later functional use.
}

