% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseHiC-class.R
\docType{class}
\name{sparseHiCdata-class}
\alias{sparseHiCdata}
\alias{sparseHiCdata-class}
\title{A class to Hi-C data from multiple samples; collection of the sparseHiCdatum
objects}
\description{
A class to Hi-C data from multiple samples; collection of the sparseHiCdatum
objects
}
\section{Slots}{

\describe{
\item{\code{HiCSamplesList}}{A list (named by samples) of lists (per sample resolutions)
of lists (Hi-C matrices)}

\item{\code{metaData}}{A data.frame of data that may be useful for}
}}

