% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseHiC-class.R
\docType{class}
\name{sparseHiCdatum-class}
\alias{sparseHiCdatum}
\alias{sparseHiCdatum-class}
\title{A class to Hi-C data from a single sample but multiple resolutions}
\description{
A class to Hi-C data from a single sample but multiple resolutions
}
\section{Slots}{

\describe{
\item{\code{sampleName}}{The name of the single sample that is the feature
of the sparseHiCdatum.}

\item{\code{resolutionNamedList}}{A list of lists of Hi-C matrices named using the
string of the resolution.}

\item{\code{metaData}}{A slot containing data corresponding to annotation data for
the particular sample.}
}}

