% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toDNAlandscapeR.R
\docType{methods}
\name{exportDNAlandscapeR}
\alias{exportDNAlandscapeR}
\alias{exportDNAlandscapeR,sparseHiCdatum-method}
\title{Export sparseHiCdatum to DNAlandscapeR files}
\usage{
exportDNAlandscapeR(dat, newSampleName = "", out.dir = "",
  splitByChr = FALSE)

\S4method{exportDNAlandscapeR}{sparseHiCdatum}(dat, newSampleName = "",
  out.dir = "", splitByChr = FALSE)
}
\arguments{
\item{dat}{A \code{sparseHiCdatum} object}

\item{newSampleName}{= "" New sample name for the object, if desired. Current
one in the object is used by default.}

\item{out.dir}{= "" The output directory to save the new tarball}

\item{splitByChr}{= FALSE  Split .rds files by chromosome for faster i/o in browser?}
}
\value{
Will produce two files in the specified out.dir
}
\description{
\code{exportDNAlandscapeR} takes a \code{sparseHiCdatum} object
and produces two files, a .rds and a .resolutions.txt file
that can be used in DNAlandscapeR (dnalandscaper.aryeelab.org). If
splitByChr is specified as TRUE, will create a folder with the sample 
name and an .rds of each chromosome for faster i/o. These split files
can only be linked via a bucket in DNAlandscapeR.
}
\examples{
# matrix.files <- paste(system.file("extdata", package = "processedHiCdata"),
#"HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_iced.matrix", sep = "/")
#bed.files <- paste(system.file("extdata", package = "processedHiCdata"),
#"HiC-Pro/hESC_Rep1/hESC_Rep1_1000000_abs.bed", sep = "/")
genomeBuild <- "hg19"
resolutions <- "1000000"
#exportDNAlandscapeR(matrix.files, bed.files, resolutions,
#sampleName = "hESC1", genomeBuild, splitByChr = TRUE)
}

